requirements.txt
asgiref==3.8.1
attrs==24.3.0
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
PyJWT==2.10.1
pyTelegramBotAPI==4.26.0
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zipp==3.21.0
kafka-python
yoomoney
